+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
   * [Enhancement] Changed websocket reconnection message from `:timeout` to
     `:re_init` for better understanding of what's happening.
+  * [Enhancement] Positive integer type.
 
 ## v0.1.6